Spreading out more files
